Blazor Rest Api continues implementations

+++ Blazor_Tutorial/EmployeeManagement.Api/Controllers/EmployeesController.cs
@@ -12,13 +12,33 @@ namespace EmployeeManagement.Api.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class EmployeesController :ControllerBase
+    public class EmployeesController : ControllerBase
     {
         private readonly IEmployeeRepository _employeeRepository;
 
         public EmployeesController(IEmployeeRepository employeeRepository)
         {
             this._employeeRepository = employeeRepository;
+        }
+
+        [HttpGet("{search}")]
+        public async Task<ActionResult<IEnumerable<Employee>>> Search(string name,Gender? gender)
+        {
+            try
+            {
+              var result= await  _employeeRepository.Search(name, gender);
+                if (result.Any())
+                {
+                    return Ok(result);
+                }
+                return NotFound();
+            }
+            catch (Exception)
+            {
+
+                return StatusCode(StatusCodes.Status500InternalServerError
+                 , "Error retrieving data from the database");
+            }
         }
 
         // GET: api/employees
@@ -118,9 +138,25 @@ namespace EmployeeManagement.Api.Controllers
     
         // DELETE: api/employees/id
         [HttpDelete("{id:int}")]
-        public async Task<ActionResult<bool>> DeleteEmployee(int id)
+        public async Task<ActionResult<Employee>> DeleteEmployee(int id)
         {
-          return await  _employeeRepository.DeleteEmployee(id);
+            try
+            {
+                var employeeToDelete = await _employeeRepository.GetEmployee(id);
+
+                if (employeeToDelete==null)
+                {
+                    return NotFound($"Employee with Id= {id} not found");
+                }
+
+                return await _employeeRepository.DeleteEmployee(id);
+            }
+            catch (Exception)
+            {
+
+                return StatusCode(StatusCodes.Status500InternalServerError
+              , "Error deleting data");
+            }
         }
     }
 }
